pyproject.toml
[project]
name = "digantara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-ninja>=1.4.3",
    "django-redis>=6.0.0",
    "gevent>=25.9.1",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
